🔥 Remove constants directory to file, start of core directory migration
