--- core/constants/galleryType.ts
@@ -1,4 +1,0 @@
-export enum GalleryType {
-    SET1,
-    SET2
-}
