--- core/constants/projectTypes.ts
@@ -1,5 +1,0 @@
-export const ProjectTypes = Object.freeze({
-    Personal: 'Personal Project',
-    Client: 'Client Work',
-    Interview: 'Tech Interview'
-});
