--- core/interfaces/gitConnectedBasics.ts
@@ -1,24 +1,0 @@
-type IGitConnectedBasics = {
-    name: string
-    label: string
-    image: string
-    email: string
-    phone: string
-    url: string
-    summary: string
-    profiles: []
-    headline: string
-    blog: string
-    yearsOfExperience: number
-    username: string
-    locationAsString: string
-    region: string
-    karma: number
-    id: string
-    followers: number
-    following: number
-    picture: string
-    website: string
-}
-
-export default IGitConnectedBasics;
