--- core/interfaces/gitConnectedEducation.ts
@@ -1,18 +1,0 @@
-type IGitConnectedEducation = {
-    institution: string
-    url: string
-    area: string
-    studyType: string
-    startDate: string
-    endDate: string
-    score: string
-    courses: [],
-    description: string
-    activities: string
-    start: { year: number, month: number },
-    end: { year: number, month: number },
-    website: string
-    gpa: string
-}
-
-export default IGitConnectedEducation;
