--- core/interfaces/gitConnectedInterest.ts
@@ -1,6 +1,0 @@
-type IGitConnectedInterest = {
-    name: string
-    keywords: []
-}
-
-export default IGitConnectedInterest;
