--- core/interfaces/gitConnectedSkill.ts
@@ -1,9 +1,0 @@
-type IGitConnectedSkill = {
-    name: string
-    level: string
-    keywords: string[]
-    rating: number
-    yearsOfExperience: number
-}
-
-export default IGitConnectedSkill;
