+++ lib/constants.ts
@@ -1,3 +1,8 @@
+export enum GalleryType {
+    SET1,
+    SET2
+}
+
 export const ProjectNames = Object.freeze({
     DigitalResume: 'Digital Resume',
     Entourage: 'Bride Entourage',
@@ -5,3 +10,9 @@ export const ProjectNames = Object.freeze({
     Kimba: 'Heyyo Kimba',
     Hundekey: 'HundekeyCreeates'
 })
+
+export const ProjectTypes = Object.freeze({
+    Personal: 'Personal Project',
+    Client: 'Client Work',
+    Interview: 'Tech Interview'
+});
